refactor: fix circular files import
