+++ app/components/inputs/autoSuggestInput/types.ts
@@ -1,0 +1,14 @@
+import { type DataListType } from "~/services/cms/components/StrapiAutoSuggestInput";
+import { type ErrorMessageProps } from "..";
+import { type FieldWidth } from "../width";
+
+export type AutoSuggestInputProps = Readonly<{
+  name: string;
+  label?: string;
+  placeholder?: string;
+  errorMessages?: ErrorMessageProps[];
+  width?: FieldWidth;
+  noSuggestionMessage?: string;
+  dataList: DataListType;
+  isDisabled: boolean;
+}>;
