+++ stories/AutoSuggestInput.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { remixContext } from "../.storybook/remixContext";
 import AutoSuggestInput from "~/components/inputs/autoSuggestInput/AutoSuggestInput";
-import type { AutoSuggestInputProps } from "~/components/inputs/autoSuggestInput/AutoSuggestInput";
+import type { AutoSuggestInputProps } from "~/components/inputs/autoSuggestInput/types";
 import { RFCFormerProvider } from ".storybook/RFCFormerProvider";
 
 function getDataListValues<T extends { dataList?: any }>(
@@ -20,7 +20,7 @@ function getDataListValues<T extends { dataList?: any }>(
   );
 }
 
-const defaulType: AutoSuggestInputProps = {
+const defaultType: AutoSuggestInputProps = {
   name: "autoSuggestInput",
   label: "Airports",
   noSuggestionMessage: "No data selected",
@@ -39,7 +39,7 @@ const meta = {
   tags: ["autodocs"],
   argTypes: {
     dataList: {
-      options: getDataListValues<AutoSuggestInputProps>(defaulType),
+      options: getDataListValues<AutoSuggestInputProps>(defaultType),
       control: { type: "select" },
     },
   },
@@ -51,7 +51,7 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {
   args: {
-    ...defaulType,
+    ...defaultType,
   },
   decorators: [
     (Story) =>
@@ -65,7 +65,7 @@ export const Default: Story = {
 
 export const WithPlaceholder: Story = {
   args: {
-    ...defaulType,
+    ...defaultType,
     placeholder: "With placeholder",
   },
   decorators: [
